backend/requirements.txt
django==5.1
djangorestframework==3.15
django-cors-headers==4.3.1
# Pillow==10.3.0
python-dotenv==1.0.1
# Pillow==10.2.0
aiohttp==3.9.3
requests==2.31.0
markdown==3.6
django-filter==24.1
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
whitenoise==6.7.0
# psycopg2-binary causes issues on some systems, only uncomment if needed
# psycopg2-binary==2.9.9
dj-database-url==2.1.0
openai==1.30.1
asgiref==3.8.1
groq==0.4.0
beautifulsoup4==4.12.3 